Swapped Events with EventsIncremental
