+++ ghc-events.cabal
@@ -33,7 +33,6 @@ library
                     bytestring >= 0.9.0,
                     array      >= 0.2 && < 0.6
   exposed-modules:  GHC.RTS.Events,
-                    GHC.RTS.EventsIncremental
                     GHC.RTS.Events.Merge
                     GHC.RTS.Events.Analysis
                     GHC.RTS.Events.Analysis.Capability
@@ -41,6 +40,7 @@ library
                     GHC.RTS.Events.Analysis.Thread
   other-modules:    GHC.RTS.EventParserUtils,
                     GHC.RTS.EventTypes
+                    GHC.RTS.EventsInternal
   include-dirs:     GHC/RTS
   extensions:	    RecordWildCards, NamedFieldPuns, BangPatterns, PatternGuards
